ccdr_appengine/main/requirements.txt
Flask==1.1.1
pandas
pylint
google-cloud-storage
